requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
httpx==0.24.1
python-multipart==0.0.6
PyPDF2==3.0.1